requirements.txt
pillow==11.3.0
requests==2.32.5
streamlit==1.50.0